# flipdot/swagger.json
{
  "definitions": {
    "Addition": {
      "type": "object",
      "properties": {
        "addition_type": {
          "type": "string",
          "enum": [
            "icon",
            "text"
          ]
        },
        "invert": {
          "type": "boolean"
        },
        "text": {
          "type": "string"
        },
        "font": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        }
      }
    },
    "DisplayData": {
      "type": "object",
      "properties": {
        "addition": {
          "$ref": "#/definitions/Addition"
        },
        "lines": {
          "type": "array",
          "maxItems": 2,
          "items": {
            "$ref": "#/definitions/TextLine"
          }
        }
      }
    },
    "TextLine": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "font": {
          "type": "string"
        },
        "invert": {
          "type": "boolean",
          "default": false
        },
        "auto_break": {
          "type": "boolean",
          "default": false
        },
        "align": {
          "type": "string",
          "enum": ["left", "center", "right"]
        }
      },
      "required": ["text"]
    }
  },
  "info": {
    "title": "Pixel Flipdot API",
    "version": "1.0"
  },
  "paths": {
...
    "/display/image": {
      "post": {
        "description": "",
        "parameters": [
          {
            "type": "integer",
            "description": "Page number, from 0 to probably 255",
            "in": "query",
            "name": "page",
            "required": true
          },
          {
            "description": "The file to upload.",
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Image should be displayed."
          }
        },
        "summary": "Display image file",
        "tags": [
          "display"
        ]
      }
    },
    "/display/text": {
      "post": {
        "description": "",
        "parameters": [
          {
            "description": "Text to be displayed",
            "in": "query",
            "name": "value",
            "required": true,
            "type": "string"
          },
          {
            "description": "Page to display on",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Font to render with. If not set using superstar.",
            "in": "query",
            "name": "font",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Text should be displayed."
          }
        },
        "summary": "Display single line of text",
        "tags": [
          "display"
        ]
      }
    },
    "/display/complex": {
      "post": {
        "description": "",
        "parameters": [
          {
            "description": "Description of data to display",
            "in": "body",
            "name": "display_data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DisplayData"
            }
          },
          {
            "description": "Page to display on",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Text should be displayed."
          },
          "400": {
            "description": "Invalid request. Observe response message."
          }
        },
        "summary": "Display complex text combination",
        "tags": [
          "display"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}